+++ flowless-library/src/main/java/flowless/Installer.java
@@ -30,6 +30,7 @@ public final class Installer {
     private KeyParceler parceler;
     private Object defaultKey;
     private Dispatcher dispatcher;
+    private Object[] globalKeys;
 
     Installer(Context baseContext, Activity activity) {
         this.baseContext = baseContext;
@@ -54,6 +55,11 @@ public final class Installer {
         return this;
     }
 
+    public Installer globalKeys(@NonNull Object... globalKeys) {
+        this.globalKeys = globalKeys;
+        return this;
+    }
+
     @NonNull
     public Context install() {
         if(InternalLifecycleIntegration.find(activity) != null) {
@@ -72,7 +78,7 @@ public final class Installer {
 
         final History defaultHistory = History.single(defState);
         final Application app = (Application) baseContext.getApplicationContext();
-        final KeyManager keyManager = new KeyManager();
+        final KeyManager keyManager = new KeyManager(globalKeys);
         final ServiceProvider serviceProvider = new ServiceProvider();
         InternalLifecycleIntegration.install(app, activity, parceler, defaultHistory, dispatcher, serviceProvider, keyManager);